Переход на использование DJS ShardingManager
И несколько оптимизаций хранения объектов в памяти
