+++ sender/exec.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 nodedir=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 cd $nodedir
+if [ "$1" = "shard" ]; then
+node shardManager.js
+echo 'Exited.'
+exit 1
+fi
 if [ "$1" = "index" ]; then
 node index.js
 echo 'Exited.'
@@ -28,4 +33,4 @@ if [ "$1" = "sendprivate" ]; then
 node sendprivate.js
 exit
 fi
-echo 'Possible args: index, multiembed, online, send, sendembed, sendprivate.'
+echo 'Possible args: shard, index, multiembed, online, send, sendembed, sendprivate.'
